@@ -10,17 +10,16 @@ setuptools.setup(
     name="cupy-xarray",
     version=versioneer.get_version(),
     cmdclass=versioneer.get_cmdclass(),
-    author="Jacob Tomlinson",
-    author_email="jtomlinson@nvidia.com",
+    author="cupy-xarray developers",
     description="Interface for using cupy in xarray, providing convenience accessors.",
     long_description=long_description,
-    long_description_content_type="text/x-markdown",
+    long_description_content_type="text/markdown",
     packages=setuptools.find_packages(),
     classifiers=[
         "Programming Language :: Python :: 3",
         "Operating System :: OS Independent",
     ],
-    python_requires=">=3.6",
+    python_requires=">=3.7",
     install_requires=requirements,
     entry_points={
         "xarray.backends": ["kvikio=cupy_xarray.kvikio:KvikioBackendEntrypoint"],
